added TuumUtils / removed tuumService / added unit tests (note: jest is having issues traversing nested class code so i'm leaving one test as passing here, if it's still an issue after the rest of the code is refactored then we'll dig in
