+++ packages/hedera-wallet-snap/packages/snap/package.json
@@ -103,6 +103,7 @@
     "prettier": "^2.8.7",
     "prettier-plugin-packagejson": "^2.4.3",
     "rimraf": "^5.0.0",
+    "ts-jest": "^29.1.2",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4",
     "webpack": "^5.78.0",
