+++ packages/hedera-wallet-snap/packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "git+https://github.com/hashgraph/hedera-metamask-snaps.git"
   },
   "source": {
-    "shasum": "7m+vEVew3y8SZ4EoXrZwqF+Lqnq8JpxGbeJBKq0i5CM=",
+    "shasum": "/ufP9PT7SMZ6j8Y4XydT0yXt9rmp1KNfKZP7ae0ulSU=",
     "location": {
       "npm": {
         "filePath": "dist/snap.js",
